requirements.txt
anyio==3.4.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
click==8.0.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
psycopg==3.0.4
psycopg2-binary==2.9.2
pydantic==1.8.2
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing_extensions==4.0.0
uvicorn==0.15.0
jwt==1.3.1
passlib==1.7.4
python-jose==3.3.0